requirements.txt
streamlit==1.28.0
pdfplumber==0.10.0
python-mammoth==1.1.0
openai==1.12.0
python-dotenv==1.0.0
